Split installer package
