+++ lib/asimov-registry/src/registry.rs
