+++ lib/asimov-registry/src/registry/error.rs
